Import of py-Glances 3.1.2 as sysutils/py-Glances
glances is a cross-platform curses-based monitoring tool which aims to
present a maximum of information in a minimum of space, ideally to fit
in a classical 80x24 terminal or higher to have additional information.
It can adapt dynamically the displayed information depending on the
terminal size. It can also work in client/server mode. Remote
monitoring could be done via terminal or web interface.

glances is written in Python and uses the psutil library to get
information from your system.