Add a compat80 package, basically containing shared libraries which exist in NetBSD 8.0, but which either doesn't exist or exist with a new major version in NetBSD 9.0_BETA.
The distributions and PLISTs produced by modifying the gencompat.sh
script (will be committed shortly), and running it over the complete
8.0 and 9.0_BETA release binaries.

This package does not attempt to separate out the X11 libraries affected
in a separate compat80-x11 package (since gencompat.sh didn't do it...).

The list of different arm variants is possibly a bit too long, but better
to err on that side than it being too short.  The package is otherwise
patterned after the compat61 package.